requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cbor2==5.6.5
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
gpiod==2.2.3
gpiodevice==0.0.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
luma.core==2.4.2
luma.lcd==2.11.0
numpy==2.2.2
openai==1.61.0
opencv-python==4.11.0.86
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyftdi==0.56.0
pyserial==3.5
python-dotenv==1.0.1
pyusb==1.3.1
requests==2.32.3
RPi.GPIO==0.7.1
smbus2==0.5.0
sniffio==1.3.1
spidev==3.6
st7735==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
